remove useless code and reorg tests
